==> 00_Intro.ipynb <==
# KBO 경기 승부 예측

### 0) 선행연구 파악

Predicting Run Production and Run Prevention in Baseball: The Impact of 
Sabermetrics

http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.457.3155&rep=rep1&type=pdf

http://cs229.stanford.edu/proj2013/JiaWongZeng-PredictingTheMajorLeagueBaseballSeason.pdf

https://towardsdatascience.com/selu-make-fnns-great-again-snn-8d61526802a9

https://www.slideshare.net/JayRyu7/baseball-prediction-model-on-tensorflow

Batting
- Batting Average (BA): number of hits divided by number of at-bats
- Slugging Percentage (SLG): number  of  bases divided by number of at-bats
- Runs Batted In (RBI): number of runs a batter’s team scores by virtue of a hit
- On-Base Percentage (OBP): how  often  a  batter reaches a base per at-bat
- Walks Drawn (WB): number  of  times  a  batter reaches a base on balls

Pitching
- Earned Run Average (ERA): earned runs per 9 innings pitched
- Hits Allowed (H): hits allowed to batters
- Bases Allowed (B): bases allowed to batters
- Strikeouts (K): strikeouts per inning pitched
- Walks Thrown (WP): number of times a pitcher allowed a base on ball

Team Stats
- Errors (E): number of fielding errors committed by team
- Past Head to Head Matchups: record  of  past games between the two teams
- Left on Base: number of batters that reach a base but never score a run
- Win Percentage (Win%): current season win percentage
