==> 03_Preprocessing.ipynb <==
import pandas as pd
import numpy as np

# Load data
matches_df = pd.read_csv("./data/matches.csv", encoding='utf-8')
matches_df = matches_df.drop(['Unnamed: 0'], axis=1)
hitter_df = pd.read_csv("./data/hitter.csv", encoding='utf-8')
pitcher_df = pd.read_csv("./data/pitcher.csv", encoding='utf-8')

matches_df.tail()

## 1) pitcher_column

- ERA: 평균자책점
- WHIP: 이닝당 출루허용률
- H: 피안타
- HR: 홈런
- SO: 삼진
- BB: 볼넷
- HBP: 사구

pitcher_df['h/g'] = pitcher_df.H / pitcher_df.G
pitcher_df['hr/g'] = pitcher_df.HR / pitcher_df.G
pitcher_df['so/g'] = pitcher_df.SO / pitcher_df.G
pitcher_df['bb/g'] = pitcher_df.BB / pitcher_df.G
pitcher_df['hbp/g'] = pitcher_df.HBP / pitcher_df.G

matches_df = matches_df.rename(columns={
    'pitcher_home': 'name'
})

target_columns = ['name', 'ERA', 'WHIP', 'h/g', 'hr/g', 'so/g', 'bb/g', 'hbp/g']

matches_df1 = matches_df.merge(pitcher_df.ix[:, target_columns], on='name', how='inner')
matches_df1.tail()

matches_df1 = matches_df1.rename(columns={
    'pitcher_away': 'name1'
})

pitcher_df1 = pitcher_df.rename(columns={
    'name': 'name1'
})

target_columns = ['name1', 'ERA', 'WHIP', 'h/g', 'hr/g', 'so/g', 'bb/g', 'hbp/g']

matches_df2 = matches_df1.merge(pitcher_df1.ix[:, target_columns], on='name1', how='inner')
matches_df2.tail()

drop_column_list = ['home_team', 'name', 'hitter_home', 
                    'away_team', 'name1', 'hitter_away']

matches_df3 = matches_df2.drop(drop_column_list, axis=1)

matches_df3.to_csv('./data/model_test.csv', encoding='utf-8')

## 2) hitter_column

- avg: 평균타율
- rbi/g: 경기당 타점
- isop: 순수장타율
- go/ao: 땅볼/뜬공

matches_df2.reset_index(inplace=True)

# hitter_df= hitter_df.replace('-', 0)
# hitter_df0 = hitter_df.ix[:, 2:].astype(float)
# hitter_df0['name'] = hitter_df['name']
# hitter_df0['team'] = hitter_df['team']

hitter_list = matches_df2.hitter_home[0].split(', ')

hitter_list

empty_array = np.zeros((10, 4))

for i, hitter in enumerate(hitter_list):
    hitter_data = np.array(hitter_df[hitter_df.name == hitter][['avg', 'isop', 'go/ao']])
    print(hitter_data)
    

# 원래는 라인업별로 적용하려 했으나, 선수 데이터가 없는 경우도 있어서... 불가능함......!!!
