==> 04_modeling.ipynb <==
import numbers
import os
import random
import pandas as pd
import numpy as np
import tensorflow as tf
from tensorflow.contrib import layers
from tensorflow.python.framework import ops, tensor_shape, tensor_util
from tensorflow.python.ops import math_ops, random_ops, array_ops
from tensorflow.python.layers import utils
from sklearn.cross_validation import train_test_split
import time
from sklearn.utils import shuffle

# load data
df = pd.read_csv("./data/model_test.csv", encoding='utf-8')
df = df.drop(['Unnamed: 0'], axis=1)
df.tail()

df = df[df.result != 0]

# define X and y
X = df.ix[:, 3:]
y = pd.get_dummies(df.result)

for feature in X.columns:
    X[feature] = (X[feature] - X[feature].mean())/X[feature].std()
    
print("Averages")
print(X.mean())

print("\n Deviations")

print(pow(X.std(), 2))

# Generate Traning and Validation Sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3)

# Convert to np arrays so that we can use with Tensorflow
X_train = np.array(X_train).astype(np.float32)
X_test  = np.array(X_test).astype(np.float32)
y_train = np.array(y_train).astype(np.float32)
y_test  = np.array(y_test).astype(np.float32)

print(np.shape(X_train), np.shape(y_train))
print(np.shape(X_test), np.shape(y_test))

# Define parameters for the model
learning_rate = 0.01
batch_size = 33
n_epochs = 10000

# Create placeholders for features and labels
X = tf.placeholder(tf.float32, [None, 14], name='X_placeholder')
Y = tf.placeholder(tf.float32, [None, 2], name='Y_placeholder')

# Create weights and bias
w = tf.Variable(tf.random_normal(shape=[14, 2], stddev=0.01), name='weights')
b = tf.Variable(tf.zeros([1, 2], name='bias'))

# Build model that returns the logits
logits = tf.matmul(X, w) + b

# Define log loss function ==> cross entropy of softmax of logits
y = tf.nn.softmax(logits)
loss = tf.reduce_mean(-tf.reduce_sum(Y * tf.log(y), reduction_indices=[1]))

optimizer = tf.train.AdamOptimizer(learning_rate).minimize(loss)

with tf.Session() as sess:
    sess.run(tf.global_variables_initializer())
    n_batches = int(len(X_train)/batch_size)
    
    for i in range(n_epochs): # train the model
        
        # shuffle X, y
        X_train, y_train = shuffle(X_train, y_train)
        total_loss = 0
        
        for j in range(n_batches):
            X_batch, Y_batch = X_train[j*batch_size:(j+1)*batch_size], y_train[j*batch_size:(j+1)*batch_size]
            _, loss_batch = sess.run([optimizer, loss], feed_dict={X: X_batch, Y:Y_batch}) 
            total_loss += loss_batch
            
        if i % 1000 == 0:
            print('Average loss epoch {0}: {1}'.format(i, total_loss/n_batches))

    print('Optimization Finished')


training_size = X_train.shape[1]
test_size = X_test.shape[1]
num_features = 14
num_labels = 2


num_hidden = 7

graph = tf.Graph()
with graph.as_default():
    tf_train_set    = tf.constant(X_train)
    tf_train_labels = tf.constant(y_train)
    tf_valid_set    = tf.constant(X_test)
 
    
    print(tf_train_set)
    print(tf_train_labels)
    
    ## Note, since there is only 1 layer there are actually no hidden layers... but if there were
    ## there would be num_hidden
    weights_1 = tf.Variable(tf.truncated_normal([num_features, num_hidden]))
    weights_2 = tf.Variable(tf.truncated_normal([num_hidden, num_labels]))
    ## tf.zeros Automaticaly adjusts rows to input data batch size
    bias_1 = tf.Variable(tf.zeros([num_hidden]))
    bias_2 = tf.Variable(tf.zeros([num_labels]))
    
    
    logits_1 = tf.matmul(tf_train_set , weights_1 ) + bias_1
    rel_1 = tf.nn.relu(logits_1)
    logits_2 = tf.matmul(rel_1, weights_2) + bias_2
    
    loss = tf.reduce_mean(tf.nn.softmax_cross_entropy_with_logits(logits=logits_2, labels=tf_train_labels))
    optimizer = tf.train.GradientDescentOptimizer(.005).minimize(loss)
    
    
    ## Training prediction
    predict_train = tf.nn.softmax(logits_2)
    
    
    
    # Validation prediction
    logits_1_val = tf.matmul(tf_valid_set, weights_1) + bias_1
    rel_1_val    = tf.nn.softmax(logits_1_val)
    logits_2_val = tf.matmul(rel_1_val, weights_2) + bias_2
    predict_valid = tf.nn.softmax(logits_2_val)

def accuracy(predictions, labels):
    return (100.0 * np.sum(np.argmax(predictions, 1) == np.argmax(labels, 1))
          / predictions.shape[0])

num_steps = 100000
with tf.Session(graph = graph) as session:
    tf.global_variables_initializer().run()
    print(loss.eval())
    test_accuracy = []
    for step in range(num_steps):
        _,l, predictions = session.run([optimizer, loss, predict_train])
        test_accuracy.append(accuracy(predict_valid.eval(), y_test))
        
        if (step % 10000 == 0 or step == num_steps-1):
#               print(predictions[3:6])
              print('Loss at step %d: %f' % (step, l))
              print('Training accuracy: %.1f%%' % accuracy(predictions, y_train[:, :]))
              print('Testing accuracy: %.1f%%' % accuracy(predict_valid.eval(), y_test))

np.max(test_accuracy)
